move concise rendering to its own module
